training progress

+++ helper.py
@@ -1,6 +1,12 @@
 import os
 import zipfile
 import requests
+import math
+import numpy as np
+from PIL import Image
+from tqdm import tqdm
+
+DEBUG = True
 
 def get_confirm_token(response):
   for key, value in response.cookies.items():
@@ -11,7 +17,8 @@ def get_confirm_token(response):
 def save_response_content(response, destination, chunk_size=32*1024):
   total_size = int(response.headers.get('content-length', 0))
   with open(destination, "wb") as f:
-    for chunk in response.iter_content(chunk_size):
+    for chunk in tqdm(response.iter_content(chunk_size), total=total_size,
+		unit='B', unit_scale=True, desc=destination):
       if chunk: # filter out keep-alive new chunks
         f.write(chunk)
 
@@ -45,9 +52,31 @@ def download_celeb_a():
   else:
     download_file_from_google_drive(drive_id, save_path)
 
-  zip_dir = ''
-  with zipfile.ZipFile(save_path) as zf:
-    zip_dir = zf.namelist()[0]
-    zf.extractall(dirpath)
-  os.remove(save_path)
-  os.rename(os.path.join(dirpath, zip_dir), os.path.join(dirpath, data_dir))
+  if not DEBUG:
+    zip_dir = ''
+    with zipfile.ZipFile(save_path) as zf:
+      zip_dir = zf.namelist()[0]
+      zf.extractall(dirpath)
+    os.remove(save_path)
+    os.rename(os.path.join(dirpath, zip_dir), os.path.join(dirpath, data_dir))
+
+def images_square_grid(images, mode='RGB'):
+    """
+    Helper function to save images as a square grid (visualization)
+    """
+    # Get maximum size for square grid of images
+    save_size = math.floor(np.sqrt(images.shape[0]))
+    # Scale to 0-255
+    images = (((images - images.min()) * 255) / (images.max() - images.min())).astype(np.uint8)
+    # Put images in a square arrangement
+    images_in_square = np.reshape(
+            images[:save_size*save_size],
+            (save_size, save_size, images.shape[1], images.shape[2], images.shape[3]))
+    # Combine images to grid image
+    new_im = Image.new(mode, (images.shape[1] * save_size, images.shape[2] * save_size))
+    for col_i, col_images in enumerate(images_in_square):
+        for image_i, image in enumerate(col_images):
+            im = Image.fromarray(image, mode)
+            new_im.paste(im, (col_i * images.shape[1], image_i * images.shape[2]))
+
+    return new_im
